move files to proper place
