--- datastore/snippets/client_with_database_test.go
@@ -1,33 +1,0 @@
-// Copyright 2024 Google LLC
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//	https://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-package datastore_snippets
-
-import (
-	"bytes"
-	"testing"
-
-	"github.com/GoogleCloudPlatform/golang-samples/internal/testutil"
-)
-
-func TestCreateClientWithDatabase(t *testing.T) {
-	test := testutil.SystemTest(t)
-	projectID := test.ProjectID
-	databaseName := "customdb"
-
-	var buf bytes.Buffer
-	err := createClientWithDatabase(&buf, projectID, databaseName)
-	if err != nil {
-		t.Fatal(err)
-	}
-}
